fix(quic): prevent size_t to int truncation in frame decode functions
Change return type from int to ssize_t for QUIC frame decode functions
to prevent potential truncation when casting size_t consumed bytes.

Changes:
- SocketQUICFrame_decode_crypto: int -> ssize_t
- SocketQUICFrame_decode_stream: int -> ssize_t
- Added sys/types.h to SocketQUICFrame.h for ssize_t definition
- Updated test files to use ssize_t for consumed variable

While QUIC frames are typically small (<1500 bytes), RFC 9000 does not
impose hard limits on reassembled stream size. Using ssize_t (standard
POSIX type for byte counts with error indication) prevents theoretical
truncation for large frames exceeding INT_MAX (2,147,483,647 bytes).

Tests: test_quic_crypto_frame_encode and test_quic_stream_frame_encode
pass with AddressSanitizer and UndefinedBehaviorSanitizer.

Closes #774

+++ src/test/test_quic_crypto_frame_encode.c
@@ -35,7 +35,7 @@ TEST (frame_crypto_encode_basic)
 
   /* Verify by parsing back */
   SocketQUICFrameCrypto_T frame;
-  int consumed = SocketQUICFrame_decode_crypto (buf, len, &frame);
+  ssize_t consumed = SocketQUICFrame_decode_crypto (buf, len, &frame);
 
   ASSERT (consumed > 0);
   ASSERT_EQ (0, frame.offset);
@@ -57,7 +57,7 @@ TEST (frame_crypto_encode_with_offset)
 
   /* Verify by parsing */
   SocketQUICFrameCrypto_T frame;
-  int consumed = SocketQUICFrame_decode_crypto (buf, len, &frame);
+  ssize_t consumed = SocketQUICFrame_decode_crypto (buf, len, &frame);
 
   ASSERT (consumed > 0);
   ASSERT_EQ (256, frame.offset);
@@ -81,7 +81,7 @@ TEST (frame_crypto_encode_large_offset)
 
   /* Verify by parsing */
   SocketQUICFrameCrypto_T frame;
-  int consumed = SocketQUICFrame_decode_crypto (buf, len, &frame);
+  ssize_t consumed = SocketQUICFrame_decode_crypto (buf, len, &frame);
 
   ASSERT (consumed > 0);
   ASSERT_EQ (large_offset, frame.offset);
@@ -102,7 +102,7 @@ TEST (frame_crypto_encode_zero_length)
 
   /* Verify by parsing */
   SocketQUICFrameCrypto_T frame;
-  int consumed = SocketQUICFrame_decode_crypto (buf, len, &frame);
+  ssize_t consumed = SocketQUICFrame_decode_crypto (buf, len, &frame);
 
   ASSERT (consumed > 0);
   ASSERT_EQ (0, frame.offset);
@@ -128,7 +128,7 @@ TEST (frame_crypto_encode_large_data)
 
   /* Verify by parsing */
   SocketQUICFrameCrypto_T frame;
-  int consumed = SocketQUICFrame_decode_crypto (buf, len, &frame);
+  ssize_t consumed = SocketQUICFrame_decode_crypto (buf, len, &frame);
 
   ASSERT (consumed > 0);
   ASSERT_EQ (100, frame.offset);
@@ -190,7 +190,7 @@ TEST (frame_crypto_encode_roundtrip_multiple)
   const uint8_t data2[] = "Second fragment";
   const uint8_t data3[] = "Third fragment";
   size_t len1, len2, len3;
-  int consumed1, consumed2, consumed3;
+  ssize_t consumed1, consumed2, consumed3;
   SocketQUICFrameCrypto_T frame;
 
   /* Encode and verify first fragment */
@@ -301,7 +301,7 @@ TEST (frame_crypto_offset_encoding)
       ASSERT (len > 0);
 
       SocketQUICFrameCrypto_T frame;
-      int consumed = SocketQUICFrame_decode_crypto (buf, len, &frame);
+      ssize_t consumed = SocketQUICFrame_decode_crypto (buf, len, &frame);
 
       ASSERT (consumed > 0);
       ASSERT_EQ (offsets[i], frame.offset);
@@ -318,7 +318,7 @@ TEST (frame_crypto_no_data_pointer_copy)
   ASSERT (len > 0);
 
   SocketQUICFrameCrypto_T frame;
-  int consumed = SocketQUICFrame_decode_crypto (buf, len, &frame);
+  ssize_t consumed = SocketQUICFrame_decode_crypto (buf, len, &frame);
   ASSERT (consumed > 0);
 
   /* Verify that frame.data points into buf, not a separate allocation */
